[feat] b-numberinput は小数にも対応しているため type は integer ではなく numeric にする

+++ app/models/quick_script/dev/calc_script.rb
@@ -11,7 +11,7 @@ module QuickScript
           {
             :label   => "左辺",
             :key     => :lhv,
-            :type    => :integer,
+            :type    => :numeric,
             :default => params[:lhv].presence.try { to_i } || 1,
           },
           {
@@ -24,7 +24,7 @@ module QuickScript
           {
             :label   => "右辺",
             :key     => :rhv,
-            :type    => :integer,
+            :type    => :numeric,
             :default => params[:rhv].presence.try { to_i } || 2,
           },
         ]
